+++ src/main/scala/draco/DomainAspect.scala
@@ -6,6 +6,8 @@ import io.circe.syntax.EncoderOps
 trait DomainAspect extends DracoType {
   val typeName: TypeName
   val elementTypeNames: Seq[String]
+  val source: TypeName
+  val target: TypeName
 }
 
 object DomainAspect extends App with DracoType {
@@ -16,7 +18,9 @@ object DomainAspect extends App with DracoType {
   implicit lazy val encoder: Encoder[DomainAspect] = Encoder.instance { x =>
     val fields = Seq(
       if (x.typeName.name.nonEmpty) Some("typeName" -> x.typeName.asJson) else None,
-      if (x.elementTypeNames.nonEmpty) Some("elementTypeNames" -> x.elementTypeNames.asJson) else None
+      if (x.elementTypeNames.nonEmpty) Some("elementTypeNames" -> x.elementTypeNames.asJson) else None,
+      if (x.source.name.nonEmpty) Some("source" -> x.source.asJson) else None,
+      if (x.target.name.nonEmpty) Some("target" -> x.target.asJson) else None
     ).flatten
     Json.obj(fields: _*)
   }
@@ -24,19 +28,25 @@ object DomainAspect extends App with DracoType {
     for {
       _typeName <- cursor.downField("typeName").as[Option[TypeName]].map(_.getOrElse(TypeName.Null))
       _elementTypeNames <- cursor.downField("elementTypeNames").as[Option[Seq[String]]].map(_.getOrElse(Seq.empty))
-    } yield DomainAspect (_typeName, _elementTypeNames)
+      _source <- cursor.downField("source").as[Option[TypeName]].map(_.getOrElse(TypeName.Null))
+      _target <- cursor.downField("target").as[Option[TypeName]].map(_.getOrElse(TypeName.Null))
+    } yield DomainAspect (_typeName, _elementTypeNames, _source, _target)
   }
 
   def apply (
     _typeName: TypeName = TypeName.Null,
-    _elementTypeNames: Seq[String] = Seq.empty
+    _elementTypeNames: Seq[String] = Seq.empty,
+    _source: TypeName = TypeName.Null,
+    _target: TypeName = TypeName.Null
   ) : DomainAspect = new DomainAspect {
     override lazy val typeName: TypeName = _typeName
     override lazy val elementTypeNames: Seq[String] = _elementTypeNames
+    override lazy val source: TypeName = _source
+    override lazy val target: TypeName = _target
     override lazy val typeDefinition: TypeDefinition = DomainAspect.typeDefinition
   }
 
   lazy val Null: DomainAspect = apply()
 
-  lazy val isEmpty: DomainAspect => Boolean = da => da.typeName.name.isEmpty && da.elementTypeNames.isEmpty
+  lazy val isEmpty: DomainAspect => Boolean = da => da.typeName.name.isEmpty && da.elementTypeNames.isEmpty && da.source.name.isEmpty && da.target.name.isEmpty
 }