+++ src/main/scala/draco/format/json/Value.scala
@@ -2,8 +2,8 @@ package draco.format.json
 
 import draco.format._
 import draco._
-import io.circe.Decoder
-import io.circe.Json
+import io.circe.{Decoder, Encoder, Json}
+import io.circe.syntax.EncoderOps
 
 trait Value extends draco.format.Value[JSON]
 
@@ -11,6 +11,20 @@ object Value extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("Value", _namePackage = Seq ("draco", "format", "json")))
   lazy val dracoType: Type[Value] = Type[Value] (typeDefinition)
   lazy val domainType: Domain[JSON] = Domain[JSON] (typeDefinition)
+
+  implicit lazy val encoder: Encoder[Value] = Encoder.instance { x =>
+    val fields = Seq(
+      Some("name" -> x.name.asJson),
+      Some("pathElements" -> x.pathElements.asJson)
+    ).flatten
+    Json.obj(fields: _*)
+  }
+  implicit lazy val decoder: Decoder[Value] = Decoder.instance { cursor =>
+    for {
+      _name <- cursor.downField("name").as[Option[String]].map(_.getOrElse(""))
+      _pathElements <- cursor.downField("pathElements").as[Option[Seq[String]]].map(_.getOrElse(Seq.empty))
+    } yield Value (_name, _pathElements)
+  }
 
   def apply (
     _name: String,
@@ -29,5 +43,6 @@ object Value extends App with DracoType {
     _name = "",
     _pathElements = Seq.empty
   )
+
 
 }